+++ src/atpx/study/__init__.py
@@ -1,8 +1,10 @@
 from .designing import Design
+from .exceptions import BlankIndexError
 from .index import LedgerIndex
 from .scaffolding import Scaffold
 
 __all__ = [
+    "BlankIndexError",
     "Design",
     "LedgerIndex",
     "Scaffold",